requirements.txt
Flask
Flask-SQLAlchemy
qrcode[pil]
python-dotenv
psycopg2-binary
